01/client/requirements.txt
grpcio
protobuf
fastapi
